+++ dirty_cat/__init__.py
@@ -12,6 +12,7 @@ from .target_encoder import TargetEncoder
 from .minhash_encoder import MinHashEncoder
 from .gap_encoder import GapEncoder
 from .super_vectorizer import SuperVectorizer
+from .pretrained_fastText import PretrainedFastText
 
 __all__ = ['SimilarityEncoder', 'TargetEncoder', 'MinHashEncoder',
-           'GapEncoder', 'SuperVectorizer']
+           'GapEncoder', 'SuperVectorizer', 'PretrainedFastText']