+++ CHANGES.rst
@@ -8,6 +8,10 @@ Version 0.2.0 is not yet releave. To install it use::
 Major changes
 -------------
 
+* **PretrainedFastText**: Added encoding of categorical variables through
+  a pretrained fastText model. FastText word embeddings capture both semantic
+  and morphological similarities.
+  
 * Bump minimum dependencies:
 
   - Python (>= 3.6)
@@ -38,7 +42,7 @@ Release 0.1.1
 
 Major changes
 -------------
-
+  
 Bug-fixes
 ---------